Implemented Triggers for enemy attacks and spawns
Enemies now have triggers that activate, when the enemies closes in on the player. then a force is applied to the enemy moving it towards the player. Direction of the force applied needs improvement.
I also tried to add triggers to the enemy spawner to make enemies spawn wen the players drives through them but couldnt get that to work quite well.