Juri branch2 (#1)
* Level Generator

* Car controller

* Free racing car assets

* Added empty camera controller script

* Enemies First steps

Made a script for the enemies, that makes the enemies chase the player.
Made a script for an empty object that instantiates enemies in a random place in a certain area.

* Added game manager for functions that need to be updated

* Updated the way it finds the car component

Updated it to find a player tag instead of an empty car tag script

* Formatting

* Changed camFollow to lateUpdate and commented line 66

* Revamped the car controller

* Implemented Triggers for enemy attacks and spawns

Enemies now have triggers that activate, when the enemies closes in on the player. then a force is applied to the enemy moving it towards the player. Direction of the force applied needs improvement.
I also tried to add triggers to the enemy spawner to make enemies spawn wen the players drives through them but couldnt get that to work quite well.

* attempt at lerp stuff

tried implementing enemy movement following the players points using nester lerps

* Added a new version of a car controller and fixed the camera system

* Convex hull script with a track generation script

* Fixed too many vectors and zero vectors from being generated from the TrackDisplacement method

* Added touch Ui controls

Touch inputs

* Added Functionality

-Main Menu
    -Start, Admin Login, Track Generator
-Encryption Added for Admin Profile
-Fixed Spawning for the car
-Added Light
-Track Generation Optimization

* Ghost added

Ghost works but the rotation is kinda fucky
theres also scripts for first experiments with the ghost where i followed the tutorial but those can be ignored (theyre the ones with ghost in the names. The ones that start with list are the important ones)

Co-authored-by: Snainner <senne010@gmail.com>
Co-authored-by: john <51299743+johntasty@users.noreply.github.com>